requirements.txt
fastapi
uvicorn[standard]
sqlalchemy[asyncio]
asyncpg
pydantic-settings
contextlib2
alembic
psycopg2-binary
Jinja2
minio
rasterio
shapely
geoalchemy2